filled out the util functions and added string defaults

+++ cyrobuf/src/cyrobuf_util.h
@@ -19,7 +19,7 @@ int64_t get_signed_varint64(const uint8_t *buffer, size_t *offset);
 int set_varint32(int32_t varint, uint8_t *buffer, size_t *offset);
 int set_varint64(int64_t varint, uint8_t *buffer, size_t *offset);
 int set_signed_varint32(int32_t varint, uint8_t *buffer, size_t *offset);
-int set_signed_varint64(int32_t varint, uint8_t *buffer, size_t *offset);
+int set_signed_varint64(int64_t varint, uint8_t *buffer, size_t *offset);
 
 #endif
